use renamed connector protocol dependency
