+++ pkg/plugin/acceptance_testing.go
@@ -24,8 +24,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1/mock"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1/mock"
 	"github.com/conduitio/conduit/pkg/foundation/assert"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/record"
