+++ pkg/plugin/builtin/v1/dispenser.go
@@ -15,7 +15,7 @@
 package builtinv1
 
 import (
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/foundation/log"
 	"github.com/conduitio/conduit/pkg/plugin"
 )
