+++ pkg/plugin/builtin/v1/dispenser_test.go
@@ -17,7 +17,7 @@ package builtinv1
 import (
 	"testing"
 
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1/mock"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1/mock"
 	"github.com/conduitio/conduit/pkg/foundation/log"
 	"github.com/conduitio/conduit/pkg/plugin"
 	"github.com/golang/mock/gomock"
