+++ pkg/plugin/builtin/v1/internal/fromplugin/record.go
@@ -17,7 +17,7 @@ package fromplugin
 import (
 	"time"
 
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/record"
 )
