+++ pkg/plugin/builtin/v1/internal/fromplugin/source.go
@@ -15,7 +15,7 @@
 package fromplugin
 
 import (
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/record"
 )
