+++ pkg/plugin/builtin/v1/internal/toplugin/source.go
@@ -15,7 +15,7 @@
 package toplugin
 
 import (
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
 	"github.com/conduitio/conduit/pkg/record"
 )
