+++ pkg/plugin/builtin/v1/internal/toplugin/specifier.go
@@ -14,7 +14,7 @@
 
 package toplugin
 
-import "github.com/conduitio/conduit-plugin-protocol/cpluginv1"
+import "github.com/conduitio/conduit-connector-protocol/cpluginv1"
 
 func SpecifierSpecifyRequest() cpluginv1.SpecifierSpecifyRequest {
 	return cpluginv1.SpecifierSpecifyRequest{}
