+++ pkg/plugin/standalone/v1/dispenser_test.go
@@ -19,9 +19,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1"
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1/mock"
-	"github.com/conduitio/conduit-plugin-protocol/cpluginv1/server"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1/mock"
+	"github.com/conduitio/conduit-connector-protocol/cpluginv1/server"
 	"github.com/conduitio/conduit/pkg/plugin"
 	"github.com/golang/mock/gomock"
 	goplugin "github.com/hashicorp/go-plugin"
