+++ pkg/plugin/standalone/v1/specifier.go
@@ -22,7 +22,7 @@ import (
 	"github.com/conduitio/conduit/pkg/plugin/standalone/v1/internal/fromproto"
 	"github.com/conduitio/conduit/pkg/plugin/standalone/v1/internal/toproto"
 	goplugin "github.com/hashicorp/go-plugin"
-	connectorv1 "go.buf.build/library/go-grpc/conduitio/conduit-plugin-protocol/connector/v1"
+	connectorv1 "go.buf.build/library/go-grpc/conduitio/conduit-connector-protocol/connector/v1"
 	"google.golang.org/grpc"
 )
